requirements.txt
dgl-cu101==0.8.1
numpy==1.18.5
numpy-indexed==0.3.5
ogb==1.3.3
pandas==1.2.3
rdflib==5.0.0
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
snorkel==0.9.8
torch==1.8.1+cu101
torch-cluster==1.5.9
torch-geometric==2.0.1
torch-scatter==2.0.6
torch-sparse==0.6.12
torch-spline-conv==1.2.1
tqdm==4.59.0